Made changes to read inputs and write outputs from/to sharepoint and to run the code in databrics
Databricks works with notebook files, so i had to create a main notebook that runs the main function for each request in the sharepoint. This code will not work on your local environment, but it works in databricks. Good to keep in mind that the sharepoint updated every dat at 6.30 am,  so any new request made will not be read until then. Maybe this is not ideal for your use case, then we need to discuss an alternative. Also, we need to schedule the job at a fixed time, ideally after the sharepoint refreshes, when you think that the output will be required.